Add CentOS support for role clamav
