+++ molecule/delegated/tests/clamav/main.py
@@ -1,17 +1,8 @@
 import pytest
 
-from .util.util import get_ansible, get_variable, get_os_role_variable
+from ..util.util import get_ansible, get_variable, get_os_role_variable
 
 testinfra_runner, testinfra_hosts = get_ansible()
-
-
-def test_pkg(host):
-    package_names = get_os_role_variable(host, "clamav_package_names")
-    assert type(package_names) is list
-
-    for package_name in package_names:
-        package = host.package(package_name)
-        assert package.is_installed
 
 
 def test_srv(host):
