+++ roles/clamav/vars/Debian.yml
@@ -9,3 +9,4 @@ clamav_localsocket: /var/run/clamav/clamd.ctl
 clamav_configuration_path: /etc/clamav/clamd.conf
 
 clamav_daemon_service_name: clamav-daemon
+clamav_freshclam_service_name: clamav-freshclam
