+++ roles/clamav/vars/RedHat.yml
@@ -9,7 +9,4 @@ clamav_localsocket: /run/clamd.scan/clamd.sock
 clamav_configuration_path: /etc/clamd.d/scan.conf
 
 clamav_daemon_service_name: clamd@scan
-
-clamav_configuration_extra:
-  - regexp: '^.*LogFile .*$'
-    line: 'LogFile /var/log/clamd.scan'
+clamav_freshclam_service_name: clamav-freshclam
